move Schema to crate::schema::Schema
